Always clear "trigger bit" from instr "control mask"
This means that the field now always contains only the length bit
(and has been renamed thus). This avoids a couple of operations
throughout FX code, at the cost of an extra one per note play
(but that's probably a tradeoff worth it), and also fixes a bug
where arpeggios could retrigger the note on each tick.