Fix incorrect processing of `D` FX for row reachability
The "pattern break" effect actually jumped to the same pattern instead of
the following one. (Or to a nonexistent pattern at the end of the song.)
This caused reachability detection to become... a little too greedy.

Additionally, pruned patterns still have references to them emitted,
causing linking errors. Something should be done about that...