Set the root freq after applying a subpattern offset
This was previously not done, but hUGEDriver actually does it!
This allows effects like portamento to act on the offset frequency,
instead of what was specified in the "main" pattern.

This is also important for "fixed mode", since it needs a way to be
cancelled somehow, and setting an offset of `+000` works just fine
for that. (Well, it didn't before this change, but it was expected to.)

At the same time, a bug was fixed, where most FX would be broken on
CH4 subpattern rows where an offset was specified. This was caused by
`de` not being preserved as pointing to the channel's note ID.